Improve keyword/keyword set form unit tests performance
