+++ src/domain/organization/organizationAuthenticationNotification/__tests__/OrganizationAuthenticationNotification.test.tsx
@@ -2,18 +2,13 @@ import { MockedResponse } from '@apollo/client/testing';
 import React from 'react';
 
 import { OrganizationDocument } from '../../../../generated/graphql';
-import {
-  fakeAuthContextValue,
-  fakeAuthenticatedAuthContextValue,
-} from '../../../../utils/mockAuthContextValue';
+import { fakeAuthenticatedAuthContextValue } from '../../../../utils/mockAuthContextValue';
 import { fakeOrganization } from '../../../../utils/mockDataUtils';
 import {
-  act,
   configure,
   CustomRenderOptions,
   render,
   screen,
-  userEvent,
   waitFor,
 } from '../../../../utils/testUtils';
 import {
@@ -84,17 +79,4 @@ test('should show notification if user has an admin organization but the id is d
 
   await screen.findByRole('heading', { name: 'Organisaatiota ei voi muokata' });
   screen.getByText('Sinulla ei ole oikeuksia muokata tätä organisaatiota.');
-});
-
-test('should start sign in process', async () => {
-  const user = userEvent.setup();
-
-  const signIn = jest.fn();
-  const authContextValue = fakeAuthContextValue({ signIn });
-  renderComponent({ authContextValue });
-
-  const signInButton = screen.getByRole('button', { name: 'kirjautua sisään' });
-  await act(async () => await user.click(signInButton));
-
-  expect(signIn).toBeCalled();
 });
